refactor(max): moved out of extended directory
BREAKING CHANGE: `max` has been moved out of `extended` folder and under `operator` folder directly
