--- src/add/operator/extended/max.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {max} from '../../../operator/extended/max';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.max = max;
-
-export var _void: void;
