+++ src/add/operator/max.ts
@@ -1,0 +1,7 @@
+import {Observable} from '../../Observable';
+import {max} from '../../operator/max';
+import {KitchenSinkOperators} from '../../Rx.KitchenSink';
+const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
+observableProto.max = max;
+
+export var _void: void;
